Prep for s19 (#1)
* Drop the material on constraints

The way that MariaDB handles constraints changed a bunch in v10.2,
and we're using 10.0. This makes what little is said in the lab write-up
obsolete, so I'm just removing that whole section for now.

* Improve formatting in `views` section

There are a lot of long queries on a single line; this breaks those up
so they're more readable, along with some other minor edits.

* Tidy up the user-defined variables section

A little formatting work, and a hint on one of the questions.

* Update a link

The MariaDB folks have clearly restructured their knowledge
library link structure. This updates one of those links; there are
many others that I should be fixing as well.